refactor: Rename RedisService to AuthTokenRedisService
1. RedisService에서 AuthTokenRedisService로 클래스명을 수정했습니다. 2. 네이버 스타일 포맷터를 적용했습니다.
